+++ Utility/Enum.lua
@@ -1,0 +1,9 @@
+require("Scenes.MenuScene")
+require("Scenes.LevelScene")
+require("Scenes.GameScene")
+
+SCENE = {
+	MENU = menu_scene,
+	GAME = game_scene,
+	LEVEL = level_scene,
+}
